+++ eggs/S/src/main/java/com/android_s/egg/AndroidSEasterEgg.kt
@@ -1,8 +1,15 @@
 package com.android_s.egg
 
 import android.app.Activity
+import android.content.ComponentName
+import android.content.Context
 import android.os.Build
+import androidx.annotation.RequiresApi
+import com.android_s.egg.neko.NekoControlsService
+import com.android_s.egg.widget.PaintChipsActivity
+import com.android_s.egg.widget.PaintChipsWidget
 import com.dede.basic.provider.BaseEasterEgg
+import com.dede.basic.provider.ComponentProvider
 import com.dede.basic.provider.EasterEgg
 import com.dede.basic.provider.EasterEggProvider
 import dagger.Module
@@ -14,7 +21,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-class AndroidSEasterEgg : EasterEggProvider {
+class AndroidSEasterEgg : EasterEggProvider, ComponentProvider {
 
     @Provides
     @IntoSet
@@ -32,6 +39,45 @@ class AndroidSEasterEgg : EasterEggProvider {
 
             override fun provideSnapshotProvider(): SnapshotProvider {
                 return SnapshotProvider()
+            }
+        }
+    }
+
+    @Provides
+    @IntoSet
+    @Singleton
+    override fun provideComponent(): ComponentProvider.Component {
+        return object : ComponentProvider.Component(
+            nameRes = R.string.s_egg_name,
+            nicknameRes = R.string.s_android_nickname,
+            apiLevel = Build.VERSION_CODES.S..Build.VERSION_CODES.S_V2
+        ) {
+            override fun isSupported(): Boolean {
+                return Build.VERSION.SDK_INT >= Build.VERSION_CODES.R
+            }
+
+            @RequiresApi(Build.VERSION_CODES.R)
+            override fun isEnabled(context: Context): Boolean {
+                val cn = ComponentName(context, NekoControlsService::class.java)
+                return cn.isEnabled(context)
+            }
+
+            @RequiresApi(Build.VERSION_CODES.R)
+            override fun setEnabled(context: Context, enable: Boolean) {
+                val cns = if (Build.VERSION.SDK_INT > Build.VERSION_CODES.R) {
+                    arrayOf(
+                        ComponentName(context, NekoControlsService::class.java),
+                        ComponentName(context, PaintChipsActivity::class.java),
+                        ComponentName(context, PaintChipsWidget::class.java)
+                    )
+                } else {
+                    arrayOf(
+                        ComponentName(context, NekoControlsService::class.java)
+                    )
+                }
+                for (cn in cns) {
+                    cn.setEnable(context, enable)
+                }
             }
         }
     }